Fix has_blob api behaviour
Removes the code line that set the body to null. Although the intention
here probably was to not have a body, the Koa framework actually seem
to alter the status code if the body is set to anything (including null).
This change ensures that the status code is set to 200 if blob exists
and 404 if not.